fix: linter errors

+++ msgblock.go
@@ -77,7 +77,7 @@ func (msg *MsgBlock) Bsvdecode(r io.Reader, pver uint32, enc MessageEncoding) er
 	// Prevent more transactions than could possibly fit into a block.
 	// It would be possible to cause memory exhaustion and panics without
 	// a sane upper bound on this count.
-	if txCount > uint64(maxTxPerBlock()) {
+	if txCount > maxTxPerBlock() {
 		str := fmt.Sprintf("too many transactions to fit into a block "+
 			"[count %d, max %d]", txCount, maxTxPerBlock())
 		return messageError("MsgBlock.Bsvdecode", str)
@@ -138,7 +138,7 @@ func (msg *MsgBlock) DeserializeTxLoc(r *bytes.Buffer) ([]TxLoc, error) {
 	// Prevent more transactions than could possibly fit into a block.
 	// It would be possible to cause memory exhaustion and panics without
 	// a sane upper bound on this count.
-	if txCount > uint64(maxTxPerBlock()) {
+	if txCount > maxTxPerBlock() {
 		str := fmt.Sprintf("too many transactions to fit into a block "+
 			"[count %d, max %d]", txCount, maxTxPerBlock())
 		return nil, messageError("MsgBlock.DeserializeTxLoc", str)
