+++ msgextendedtx.go
@@ -184,7 +184,7 @@ func (msg *MsgExtendedTx) Bsvdecode(r io.Reader, pver uint32, enc MessageEncodin
 	// Prevent more input transactions than could possibly fit into a
 	// message.  It would be possible to cause memory exhaustion and panics
 	// without a sane upper bound on this count.
-	if count > uint64(maxTxInPerMessage()) {
+	if count > maxTxInPerMessage() {
 		str := fmt.Sprintf("too many input transactions to fit into "+
 			"max message size [count %d, max %d]", count, maxTxInPerMessage())
 		return messageError("MsgTx.Bsvdecode", str)
@@ -245,7 +245,7 @@ func (msg *MsgExtendedTx) Bsvdecode(r io.Reader, pver uint32, enc MessageEncodin
 	// Prevent more output transactions than could possibly fit into a
 	// message.  It would be possible to cause memory exhaustion and panics
 	// without a sane upper bound on this count.
-	if count > uint64(maxTxOutPerMessage()) {
+	if count > maxTxOutPerMessage() {
 		returnScriptBuffers()
 
 		str := fmt.Sprintf("too many output transactions to fit into "+
